{{cookiecutter.project_name}}/core/pyproject.toml
[tool.poetry]
name = "innohack-backend"
version = "0.1.0"
description = "Server side for innohack project"
authors = ["Igor Beschastnov <beschastnovigor@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
gunicorn = "^20.0.4"
django = "^3.1.2"
psycopg2-binary = "^2.8.6"
django-environ = "^0.4.5"
djangorestframework = "^3.12.1"
django-filter = "^2.4.0"
django-cors-headers = "^3.5.0"
drf-yasg = "^1.20.0"
django-extensions = "^3.0.9"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
isort = "^5.6.4"
unify = "^0.5"
mypy = "^0.790"
pylint = "^2.6.0"
django-webpack-loader = "^0.7.0"

[build-system]
requires = ["poetry>=1.0.5"]
build-backend = "poetry.masonry.api"
